# feat: improve std.reverse performance by 73%