--- a/builtin-benchmarks/reverse.jsonnet
+++ b/builtin-benchmarks/reverse.jsonnet
@@ -1,0 +1,5 @@
+{
+    foo: [
+        std.reverse(std.range(0, 5000)) for i in std.range(0,100)
+    ],
+}