--- a/builtins_benchmark_test.go
+++ b/builtins_benchmark_test.go
@@ -2,6 +2,7 @@ package jsonnet
 
 import (
 	"flag"
+	"fmt"
 	"os"
 	"os/exec"
 	"testing"
@@ -15,9 +16,9 @@ func init() {
 	flag.BoolVar(&outputPassthru, "outputPassthru", false, "Pass stdout/err from jsonnet")
 }
 
-func Benchmark_Builtin_substr(b *testing.B) {
+func RunBenchmark(b *testing.B, name string) {
 	for n := 0; n < b.N; n++ {
-		cmd := exec.Command(jsonnetPath, "./builtin-benchmarks/substr.jsonnet")
+		cmd := exec.Command(jsonnetPath, fmt.Sprintf("./builtin-benchmarks/%s.jsonnet", name))
 		if outputPassthru {
 			cmd.Stdout = os.Stdout
 			cmd.Stderr = os.Stderr
@@ -28,4 +29,12 @@ func Benchmark_Builtin_substr(b *testing.B) {
 			b.Fail()
 		}
 	}
+}
+
+func Benchmark_Builtin_substr(b *testing.B) {
+	RunBenchmark(b, "substr")
+}
+
+func Benchmark_Builtin_reverse(b *testing.B) {
+	RunBenchmark(b, "reverse")
 }