--- a/testdata/builtinReverse.jsonnet
+++ b/testdata/builtinReverse.jsonnet
@@ -1,0 +1,1 @@
+std.reverse([1, 2])