--- a/testdata/builtinReverse_empty.jsonnet
+++ b/testdata/builtinReverse_empty.jsonnet
@@ -1,0 +1,1 @@
+std.reverse([])