--- a/testdata/builtinReverse_many.golden
+++ b/testdata/builtinReverse_many.golden
@@ -1,0 +1,7 @@
+[
+   "tester",
+   "is",
+   "name",
+   "my",
+   "hello"
+]