--- a/testdata/builtinReverse_many.jsonnet
+++ b/testdata/builtinReverse_many.jsonnet
@@ -1,0 +1,1 @@
+std.reverse(["hello", "my", "name", "is", "tester"])