--- a/testdata/builtinReverse_not_array.golden
+++ b/testdata/builtinReverse_not_array.golden
@@ -1,0 +1,10 @@
+RUNTIME ERROR: Unexpected type string, expected array
+-------------------------------------------------
+	testdata/builtinReverse_not_array:1:1-20	builtin function <reverse>
+
+std.reverse("asdf")
+
+-------------------------------------------------
+	During evaluation	
+
+