--- a/testdata/builtinReverse_not_array.jsonnet
+++ b/testdata/builtinReverse_not_array.jsonnet
@@ -1,0 +1,1 @@
+std.reverse("asdf")