--- a/testdata/builtinReverse_single.jsonnet
+++ b/testdata/builtinReverse_single.jsonnet
@@ -1,0 +1,1 @@
+std.reverse(["hello"])