… testdata/builtinReverse.golden
… testdata/builtinReverse_empty.golden
… testdata/builtinReverse_single.golden