Implements std.reverse in native Go, improving performance

```
benchmark                        old ns/op     new ns/op     delta
Benchmark_Builtin_reverse-16     869191619     231309458     -73.39%
```

### Notes

~This PR contains a commit from another pending PR, and will be rebased once #343 is merged.~