@ghostsquad 2020-01-26
This has been rebased, and is ready for review/merge @sbarzowski